+++ NEWS.md
@@ -1,5 +1,7 @@
 # FSA 0.7.1 ongoing
 * `hist.formula()`: Modified.  Fixed so that `ymax=` also sets the y-axis limit when only one histogram is made (it was previously ignored).
+* `iHndlFormula()`: Modified.  Added `na.action=NULL` to `model.frame()` so that NA values will not be omitted.
+* `psdAdd()`: Modified.  Fixed bug that never tested if all lengths were `NA`.  Required change to `iHndlFormula`.
 
 # FSA 0.7.0 Jul15
 * Fixed description to be in title case.